Added support for HierarchicalPoolInterface
