+++ composer.json
@@ -28,11 +28,12 @@
     ],
     "require":
     {
-        "php":                  "^5.5|^7.0",
-        "psr/cache":            "1.0.0",
-        "cache/adapter-common": "^0.1.2",
-        "cache/taggable-cache": "^0.2",
-        "predis/predis":        "^1.0"
+        "php":                      "^5.5|^7.0",
+        "psr/cache":                "1.0.0",
+        "cache/adapter-common":     "^0.1.2",
+        "cache/taggable-cache":     "^0.3",
+        "cache/hierarchical-cache": "^0.1",
+        "predis/predis":            "^1.0"
     },
     "require-dev":
     {
